Move components into their own files
